Add Transform component.
Now we can create graphical components that are not connected to a Box2D body (for example input indicators; entities of only graphical and transform components). All entities with a Physical component must have a Transform component (values are being updated in Physics system), and all Graphical components must have a Transform component (to be able to draw it with correct position/scale/rotation and if it visible or not).